cdk/lambda/sessions/requirements.txt
strands-agents==1.11.0
boto3==1.40.24
aws-lambda-powertools==3.19.0
pydantic>=2.0.0